refactor(cloud-function): rename middlewares {fooBar => foo-bar}
