+++ cloud-function/src/middlewares/redirect-fundamental.ts
