+++ cloud-function/src/middlewares/redirect-leading-slash.ts
