+++ cloud-function/src/middlewares/redirect-locale.ts
