+++ cloud-function/src/middlewares/redirect-moved-pages.ts
