+++ cloud-function/src/middlewares/redirect-trailing-slash.ts
