+++ cloud-function/src/middlewares/require-origin.ts
