+++ cloud-function/src/middlewares/resolve-index-html.ts
